TS builder refactor

+++ script.ts/build.Typescript.mts
@@ -1,4 +1,4 @@
-import { exec } from 'child_process';
+import { exec, ExecException } from 'child_process';
 import fs from 'fs-extra';
 import path from 'path';
 
@@ -6,8 +6,11 @@ type PathString = string;
 type TsConfig = { extends?: string; include?: string[]; compilerOptions: TsConfigCompilerOptions };
 type TsConfigCompilerOptions = { rootDir?: string };
 
+/**
+ * Template path names.
+ */
 const Path = {
-  builderDir: '.builder',
+  tmpBuilderDir: '.builder',
   filename: {
     code: 'tsconfig.code.json',
     types: 'tsconfig.types.json',
@@ -23,16 +26,21 @@ export const Typescript = {
   /**
    * Run the [tsc] typescript compiler
    */
-  tsc(dir: PathString, tsconfig: PathString) {
-    type R = { ok: boolean; stdout: string; stderr: string };
+  tsc(dir: PathString, tsconfig: PathString, options: { silent?: boolean } = {}) {
+    type R = { ok: boolean; code: number; stdout: string; stderr: string; error?: ExecException };
     return new Promise<R>(async (resolve, reject) => {
       const tscompiler = path.resolve('node_modules/typescript/bin/tsc');
       const cmd = `${tscompiler} --project "${path.join(dir, tsconfig)}"`;
-      exec(cmd, (error, stdout, stderr) => {
-        const ok = !Boolean(stderr);
-        if (stdout) console.log(stdout);
-        if (stderr) console.error(stderr);
-        return error ? reject(error) : resolve({ ok, stdout, stderr });
+      exec(cmd, (err, stdout, stderr) => {
+        const ok = !Boolean(stderr || err);
+        const code = typeof err?.code === 'number' ? err.code : ok ? 0 : 1;
+        const error = err || undefined;
+        if (!options.silent) {
+          if (stdout) console.log(`\n${stdout}`);
+          if (stderr) console.error(`\n${stderr}`);
+          if (error) console.error(`[command failed]:\n${cmd}\n`);
+        }
+        resolve({ ok, code, stdout, stderr, error });
       });
     });
   },
@@ -40,42 +48,50 @@ export const Typescript = {
   /**
    * Complete build
    */
-  async build(dir: PathString) {
-    await Typescript.copyConfigs(dir, { clear: true });
-    await Typescript.buildCode(dir);
-    await Typescript.buildTypes(dir);
-    await fs.remove(path.join(dir, Path.builderDir));
+  async build(rootDir: PathString, options: { exitOnError?: boolean } = {}) {
+    await Typescript.copyTsConfigFiles(rootDir, { clear: true });
+    await Typescript.buildCode(rootDir, options);
+    await Typescript.buildTypes(rootDir, options);
+    await fs.remove(path.join(rootDir, Path.tmpBuilderDir));
   },
 
   /**
    * Build ESM code.
    */
-  async buildCode(dir: PathString) {
-    const tsconfig = path.join(Path.builderDir, Path.filename.code);
-    await Typescript.tsc(dir, tsconfig);
+  async buildCode(rootDir: PathString, options: { exitOnError?: boolean } = {}) {
+    const tsconfig = path.join(Path.tmpBuilderDir, Path.filename.code);
+    const res = await Typescript.tsc(rootDir, tsconfig);
+    if (!res.ok && options.exitOnError) process.exit(res.code);
+    return res;
   },
 
   /**
    * Build type definitions (.d.ts)
    */
-  async buildTypes(dir: PathString) {
-    const tsconfig = path.join(Path.builderDir, Path.filename.types);
-    await Typescript.tsc(dir, tsconfig);
+  async buildTypes(rootDir: PathString, options: { exitOnError?: boolean } = {}) {
+    const tsconfig = path.join(Path.tmpBuilderDir, Path.filename.types);
+    const res = await Typescript.tsc(rootDir, tsconfig);
+    if (!res.ok && options.exitOnError) process.exit(res.code);
 
-    const source = path.join(dir, Path.builderDir, 'types/src');
-    const target = path.join(dir, 'types');
-    await fs.remove(target);
-    await fs.move(source, target);
-    await fs.remove(path.join(dir, Path.builderDir, 'types'));
+    // Move the child "src/" folder up into the root "types/" folder.
+    if (res.ok) {
+      const source = path.join(rootDir, Path.tmpBuilderDir, 'types/src');
+      const target = path.join(rootDir, 'types');
+      await fs.remove(target);
+      await fs.move(source, target);
+      await fs.remove(path.join(rootDir, Path.tmpBuilderDir, 'types'));
+    }
+
+    return res;
   },
 
   /**
    * Copy the [tsconfig] json files to the target directory.
    */
-  async copyConfigs(dir: PathString, options: { clear?: boolean } = {}) {
-    dir = path.resolve(dir);
+  async copyTsConfigFiles(rootDir: PathString, options: { clear?: boolean } = {}) {
+    rootDir = path.resolve(rootDir);
     const sourceDir = path.resolve('./template');
-    const targetDir = path.join(dir, Path.builderDir);
+    const targetDir = path.join(rootDir, Path.tmpBuilderDir);
     if (options.clear) await fs.remove(targetDir);
     await fs.ensureDir(targetDir);
 
@@ -89,11 +105,11 @@ export const Typescript = {
 
     await copy(Path.filename.code, (tsconfig) => {
       tsconfig.extends = path.resolve('./tsconfig.json');
-      tsconfig.compilerOptions.rootDir = dir;
+      tsconfig.compilerOptions.rootDir = rootDir;
     });
 
     await copy(Path.filename.types, (tsconfig) => {
-      tsconfig.compilerOptions.rootDir = dir;
+      tsconfig.compilerOptions.rootDir = rootDir;
     });
   },
 };
